added unit tests for services and model

+++ src/Interface/Service/SalesTaxServiceInterface.php
@@ -6,5 +6,5 @@ use Src\Interface\Model\CustomerInterface;
 
 interface SalesTaxServiceInterface
 {
-    public function calculate(float $amount, CustomerInterface $customer): int;
+    public function calculate(float $amount, CustomerInterface $customer): float;
 }
